app/requirements.txt
gensim
python-Levenshtein
requests
python-dotenv
esprima
numpy
scipy
google-cloud-datastore
antlr4-python3-runtime
